Set 'multiline' matching on by default
This change prefixes every pattern with '(?m)' which enables Oniguruma's
"multiline" mode. The only effect of this is that the '.' atom will
match newline characters.

This should solve a common problem that newbies use the multiline filter
and then try to grok on it, getting previously burned by grok not
allowing '.*' to match across newlines.